読み込み専用モード時に、書き込み・変更が発生するファイル操作が出来ないように修正

+++ mtnd.c
@@ -453,6 +453,15 @@ static void mtnd_truncate_process(MTNTASK *kt)
   mtndata_get_string(path, &(kt->recv));
   mtndata_get_int(&offset, &(kt->recv), sizeof(offset));
   mtnd_fix_path(path, kt->path);
+  if(ctx->rdonly){
+    if(lstat(kt->path, &(kt->stat)) != -1){
+      errno = EROFS;
+      kt->send.head.type = MTNCMD_ERROR;
+      mtndata_set_int(&errno, &(kt->send), sizeof(errno));
+      mtnlogger(mtn, 0, "[error] %s: %s path=%s offset=%llu\n", __func__, strerror(errno), kt->path, offset);
+    }
+    goto ERROR;
+  }
   if(truncate(kt->path, offset) == -1){
     if(errno != ENOENT){
       kt->send.head.type = MTNCMD_ERROR;
@@ -460,6 +469,7 @@ static void mtnd_truncate_process(MTNTASK *kt)
       mtnlogger(mtn, 0, "[error] %s: %s path=%s offset=%llu\n", __func__, strerror(errno), kt->path, offset);
     }
   }
+ERROR:
   mtnlogger(mtn, 9, "[debug] %s: OUT\n", __func__);
 }
 
@@ -478,12 +488,21 @@ static void mtnd_mkdir_process(MTNTASK *kt)
   mtndata_get_int(&gid, &(kt->recv), sizeof(gid));
   mtnd_fix_path(buff, kt->path);
   kt->send.head.type = MTNCMD_SUCCESS;
-  if(mkdir_ex(kt->path) == -1){
-    mtnlogger(mtn, 0, "[error] %s: %s %s\n", __func__, strerror(errno), kt->path);
+  if(ctx->rdonly){
+    errno = EROFS;
     kt->send.head.type = MTNCMD_ERROR;
     mtndata_set_int(&errno, &(kt->send), sizeof(errno));
+    mtnlogger(mtn, 0, "[error] %s: %s %s\n", __func__, strerror(errno), kt->path);
+    goto ERROR;
+  }
+  if(mkdir_ex(kt->path) == -1){
+    kt->send.head.type = MTNCMD_ERROR;
+    mtndata_set_int(&errno, &(kt->send), sizeof(errno));
+    mtnlogger(mtn, 0, "[error] %s: %s %s\n", __func__, strerror(errno), kt->path);
+    goto ERROR;
   }
   chown(kt->path, uid, gid);
+ERROR:
   mtnlogger(mtn, 9, "[debug] %s: OUT\n", __func__);
 }
 
@@ -499,20 +518,28 @@ static void mtnd_rm_process(MTNTASK *kt)
   mtnd_fix_path(buff, kt->path);
   kt->send.head.type = MTNCMD_SUCCESS;
   if(lstat(kt->path, &(kt->stat)) != -1){
+    if(ctx->rdonly){
+      errno = EROFS;
+      kt->send.head.type = MTNCMD_ERROR;
+      mtndata_set_int(&errno, &(kt->send), sizeof(errno));
+      mtnlogger(mtn, 0, "[error] %s: %s %s\n", __func__, strerror(errno), kt->path);
+      goto ERROR;
+    }
     if(S_ISDIR(kt->stat.st_mode)){
       if(rmdir(kt->path) == -1){
-        mtnlogger(mtn, 0, "[error] %s: %s %s\n", __func__, strerror(errno), kt->path);
         kt->send.head.type = MTNCMD_ERROR;
         mtndata_set_int(&errno, &(kt->send), sizeof(errno));
+        mtnlogger(mtn, 0, "[error] %s: %s %s\n", __func__, strerror(errno), kt->path);
       }
     }else{
       if(unlink(kt->path) == -1){
-        mtnlogger(mtn, 0, "[error] %s: %s %s\n", __func__, strerror(errno), kt->path);
         kt->send.head.type = MTNCMD_ERROR;
         mtndata_set_int(&errno, &(kt->send), sizeof(errno));
+        mtnlogger(mtn, 0, "[error] %s: %s %s\n", __func__, strerror(errno), kt->path);
       }
     }
   }
+ERROR:
   mtnlogger(mtn, 9, "[debug] %s: OUT\n", __func__);
 }
 
@@ -530,13 +557,23 @@ static void mtnd_rename_process(MTNTASK *kt)
   mtnd_fix_path(obuff, kt->path);
   mtnd_fix_path(nbuff, NULL);
   kt->send.head.type = MTNCMD_SUCCESS;
-  if(rename(kt->path, nbuff) == -1){
-    if(errno != ENOENT){
-      mtnlogger(mtn, 0, "[error] %s: %s %s -> %s\n", __func__, strerror(errno), obuff, nbuff);
+  if(ctx->rdonly){
+    if(lstat(kt->path, &(kt->stat)) != -1){
+      errno = EROFS;
       kt->send.head.type = MTNCMD_ERROR;
       mtndata_set_int(&errno, &(kt->send), sizeof(errno));
+      mtnlogger(mtn, 0, "[error] %s: %s %s -> %s\n", __func__, strerror(errno), obuff, nbuff);
+    }
+    goto ERROR;
+  }
+  if(rename(kt->path, nbuff) == -1){
+    if(errno != ENOENT){
+      kt->send.head.type = MTNCMD_ERROR;
+      mtndata_set_int(&errno, &(kt->send), sizeof(errno));
+      mtnlogger(mtn, 0, "[error] %s: %s %s -> %s\n", __func__, strerror(errno), obuff, nbuff);
     }
   }
+ERROR:
   mtnlogger(mtn, 9, "[debug] %s: OUT\n", __func__);
 }
 
@@ -553,11 +590,19 @@ static void mtnd_symlink_process(MTNTASK *kt)
   mtndata_get_string(newpath, &(kt->recv));
   mtnd_fix_path(newpath, NULL);
   kt->send.head.type = MTNCMD_SUCCESS;
-  if(symlink(oldpath, newpath) == -1){
-    mtnlogger(mtn, 0, "[error] %s: %s %s -> %s\n", __func__, strerror(errno), oldpath, newpath);
+  if(ctx->rdonly){
+    errno = EROFS;
     kt->send.head.type = MTNCMD_ERROR;
     mtndata_set_int(&errno, &(kt->send), sizeof(errno));
+    mtnlogger(mtn, 0, "[error] %s: %s %s -> %s\n", __func__, strerror(errno), oldpath, newpath);
+    goto ERROR;
   }
+  if(symlink(oldpath, newpath) == -1){
+    kt->send.head.type = MTNCMD_ERROR;
+    mtndata_set_int(&errno, &(kt->send), sizeof(errno));
+    mtnlogger(mtn, 0, "[error] %s: %s %s -> %s\n", __func__, strerror(errno), oldpath, newpath);
+  }
+ERROR:
   mtnlogger(mtn, 9, "[debug] %s: OUT\n", __func__);
 }
 
@@ -599,13 +644,23 @@ static void mtnd_chmod_process(MTNTASK *kt)
   mtndata_get_int(&mode, &(kt->recv), sizeof(mode));
   mtnd_fix_path(path, NULL);
   kt->send.head.type = MTNCMD_SUCCESS;
-  if(chmod(path, mode) == -1){
-    if(errno != ENOENT){
-      mtnlogger(mtn, 0, "[error] %s: %s %s\n", __func__, strerror(errno), path);
+  if(ctx->rdonly){
+    if(lstat(path, &(kt->stat)) != -1){
+      errno = EROFS;
       kt->send.head.type = MTNCMD_ERROR;
       mtndata_set_int(&errno, &(kt->send), sizeof(errno));
+      mtnlogger(mtn, 0, "[error] %s: %s %s\n", __func__, strerror(errno), path);
+    }
+    goto ERROR;
+  }
+  if(chmod(path, mode) == -1){
+    if(errno != ENOENT){
+      kt->send.head.type = MTNCMD_ERROR;
+      mtndata_set_int(&errno, &(kt->send), sizeof(errno));
+      mtnlogger(mtn, 0, "[error] %s: %s %s\n", __func__, strerror(errno), path);
     }
   }
+ERROR:
   mtnlogger(mtn, 7, "[debug] %s: OUT\n", __func__);
 }
 
@@ -624,13 +679,23 @@ static void mtnd_chown_process(MTNTASK *kt)
   mtndata_get_int(&gid, &(kt->recv), sizeof(gid));
   mtnd_fix_path(path, NULL);
   kt->send.head.type = MTNCMD_SUCCESS;
-  if(chown(path, uid, gid) == -1){
-    if(errno != ENOENT){
-      mtnlogger(mtn, 0, "[error] %s: %s %s\n", __func__, strerror(errno), path);
+  if(ctx->rdonly){
+    if(lstat(path, &(kt->stat)) != -1){
+      errno = EROFS;
       kt->send.head.type = MTNCMD_ERROR;
       mtndata_set_int(&errno, &(kt->send), sizeof(errno));
+      mtnlogger(mtn, 0, "[error] %s: %s %s\n", __func__, strerror(errno), path);
+    }
+    goto ERROR;
+  }
+  if(chown(path, uid, gid) == -1){
+    if(errno != ENOENT){
+      kt->send.head.type = MTNCMD_ERROR;
+      mtndata_set_int(&errno, &(kt->send), sizeof(errno));
+      mtnlogger(mtn, 0, "[error] %s: %s %s\n", __func__, strerror(errno), path);
     }
   }
+ERROR:
   mtnlogger(mtn, 7, "[debug] %s: OUT\n", __func__);
 }
 
@@ -648,13 +713,23 @@ static void mtnd_utime_process(MTNTASK *kt)
   mtndata_get_int(&(ut.modtime), &(kt->recv), sizeof(ut.modtime));
   mtnd_fix_path(path, NULL);
   kt->send.head.type = MTNCMD_SUCCESS;
-  if(utime(path, &ut) == -1){
-    if(errno != ENOENT){
-      mtnlogger(mtn, 0, "[error] %s: %s %s\n", __func__, strerror(errno), path);
+  if(ctx->rdonly){
+    if(lstat(path, &(kt->stat)) != -1){
+      errno = EROFS;
       kt->send.head.type = MTNCMD_ERROR;
       mtndata_set_int(&errno, &(kt->send), sizeof(errno));
+      mtnlogger(mtn, 0, "[error] %s: %s %s\n", __func__, strerror(errno), path);
+    }
+    goto ERROR;
+  }
+  if(utime(path, &ut) == -1){
+    if(errno != ENOENT){
+      kt->send.head.type = MTNCMD_ERROR;
+      mtndata_set_int(&errno, &(kt->send), sizeof(errno));
+      mtnlogger(mtn, 0, "[error] %s: %s %s\n", __func__, strerror(errno), path);
     }
   }
+ERROR:
   mtnlogger(mtn, 7, "[debug] %s: OUT\n", __func__);
 }
